"rscli auto-commit"

+++ internal/storage/postgres/user_queries/querier.go
@@ -11,7 +11,7 @@ import (
 
 type Querier interface {
 	Add(ctx context.Context, arg AddParams) error
-	GetByTelegramId(ctx context.Context, telegramID sql.NullInt64) (GetByTelegramIdRow, error)
+	GetByTelegramId(ctx context.Context, telegramID sql.NullInt32) (GetByTelegramIdRow, error)
 	GetByTelegramName(ctx context.Context, username string) (GetByTelegramNameRow, error)
 	GetPassByUsername(ctx context.Context, username string) (string, error)
 	ListStatistics(ctx context.Context) ([]UserStatistic, error)
